requirements.txt
streamlit==1.25.0
pandas==1.5.3
beautifulsoup4==4.10.0
google-generativeai==0.0.11
py-pdf2==1.26.0
docx2txt==0.0.15
joblib==1.2.0
requests==2.32.3